add publish task
added a new task to allow publishing the assets directly to a location outside the frontend repo.
Again this is useful for backend developers which want to publish the assets directly to their "web-root" (or whatever it is called in the respective technology).
The publish task do nothing, except you set the `--publishDest` parameter on the command line.